requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# ── Database ────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
databases[asyncpg]==0.9.0
sqlalchemy==2.0.30
asyncpg==0.29.0

# ── dbt ─────────────────────────────────────────────────────────────────────
dbt-core==1.8.2
dbt-postgres==1.8.2

# ── Airflow extras (installed inside Airflow image via pip) ──────────────────
apache-airflow-providers-postgres==5.10.2

# ── LangChain + Ollama ───────────────────────────────────────────────────────
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
ollama==0.2.0

# ── Data processing ──────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0
